backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "UNSW capstone-project-25t3-9900-h18f-cake"
authors = [
    {name = "jiahao li",email = "z5501964@ad.unsw.edu.au"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (==0.121.0)",
    "uvicorn (==0.38.0)",
    "pydantic-settings (==2.11.0)",
    "pydantic (==2.12.3)",
    "passlib (==1.7.4)",
    "python-multipart (==0.0.20)",
    "bcrypt (==5.0.0)",
    "pyjwt (>=2.8.0)",
    "requests (==2.32.5)",
    "numpy (==1.26.4)",
    "openai (>=1.40.0)",
    "pymupdf (==1.24.9)",
    "annotated-doc (==0.0.3)",
    "annotated-types (==0.7.0)",
    "anyio (==4.11.0)",
    "certifi (==2025.10.5)",
    "cffi (==2.0.0)",
    "chardet (==5.2.0)",
    "charset-normalizer (==3.4.4)",
    "click (==8.3.0)",
    "colorlog (==6.10.1)",
    "contourpy (==1.3.3)",
    "cryptography (==46.0.3)",
    "cycler (==0.12.1)",
    "distro (==1.9.0)",
    "dnspython (==2.8.0)",
    "ecdsa (==0.19.1)",
    "email-validator (==2.3.0)",
    "faiss-cpu (==1.8.0)",
    "filelock (==3.20.0)",
    "fonttools (==4.60.1)",
    "fsspec (==2025.10.0)",
    "future (==1.0.0)",
    "h11 (==0.16.0)",
    "hf-xet (==1.2.0)",
    "httpcore (==1.0.9)",
    "httpx (==0.28.1)",
    "huggingface-hub (==0.36.0)",
    "idna (==3.11)",
    "imagesize (==1.4.1)",
    "jinja2 (==3.1.6)",
    "jiter (==0.11.1)",
    "joblib (==1.5.2)",
    "keybert (>=0.8.4)",
    "kiwisolver (==1.4.9)",
    "lxml (==6.0.2)",
    "markdown-it-py (==4.0.0)",
    "markupsafe (==3.0.3)",
    "matplotlib (==3.9.2)",
    "mdurl (==0.1.2)",
    "modelscope (==1.31.0)",
    "mpmath (==1.3.0)",
    "networkx (==3.5)",
    "nltk (==3.9.1)",
    "opencv-contrib-python (==4.10.0.84)",
    "opencv-python (==4.11.0.86)",
    "opencv-python-headless (==4.10.0.84)",
    "opt-einsum (==3.3.0)",
    "packaging (==25.0)",
    "pandas (==2.2.2)",
    "pdfminer-six (==20240706)",
    "pillow (==10.3.0)",
    "prettytable (==3.16.0)",
    "protobuf (==6.33.0)",
    "psutil (==7.1.3)",
    "py-cpuinfo (==9.0.0)",
    "pyasn1 (==0.6.1)",
    "pyclipper (==1.3.0.post6)",
    "pycparser (==2.23)",
    "pycryptodome (==3.23.0)",
    "pydantic-core (==2.41.4)",
    "pygments (==2.19.2)",
    "pymupdfb (==1.24.9)",
    "pyparsing (==3.2.5)",
    "pypdf (==4.3.1)",
    "pypdfium2 (==5.0.0)",
    "pytesseract (==0.3.10)",
    "python-bidi (==0.6.7)",
    "python-dateutil (==2.9.0.post0)",
    "python-docx (==1.1.2)",
    "python-dotenv (==1.2.1)",
    "python-jose (==3.5.0)",
    "pytz (==2025.2)",
    "pyyaml (==6.0.2)",
    "regex (>=2023.10.3)",
    "rich (==14.2.0)",
    "ruamel-yaml (==0.18.16)",
    "ruamel-yaml-clib (==0.2.14)",
    "safetensors (==0.6.2)",
    "scikit-learn (==1.5.1)",
    "scipy (==1.16.3)",
    "sentence-transformers (==2.7.0)",
    "shapely (==2.1.2)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.44)",
    "starlette (==0.49.3)",
    "sympy (==1.14.0)",
    "threadpoolctl (==3.6.0)",
    "tokenizers (==0.22.1)",
    "tqdm (==4.66.4)",
    "transformers (==4.57.1)",
    "typing-inspection (==0.4.2)",
    "typing-extensions (==4.15.0)",
    "tzdata (==2025.2)",
    "ujson (==5.11.0)",
    "urllib3 (==2.5.0)",
    "wcwidth (==0.2.14)"
]

[tool.poetry]
name = "app"
version = "0.1.0"
description = "UNSW capstone-project-25t3-9900-h18f-cake"
authors = ["jiahao li <z5501964@ad.unsw.edu.au>"]
packages = [
  { include = "app" }
]

[[tool.poetry.source]]
name = "paddle-stable-cpu"
url = "https://www.paddlepaddle.org.cn/packages/stable/cpu/"
priority = "explicit"


[tool.poetry.dependencies]
python = ">=3.12"
paddlepaddle = {version = "3.2.1", source = "paddle-stable-cpu"}
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "isort (>=6.1.0,<7.0.0)",
    "alembic (==1.13.2)",
    "pytest (>=8.3.0,<9.0.0)",
    "pytest-cov (>=5.0.0,<6.0.0)"
]
